requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
passlib[bcrypt]
pyjwt
pdfminer.six
python-docx
